Add 5 methods for interacting with movie data
getSearch - will initiate a call to the movie database api based on the
search term provided in the client request.  Once a list of movies has
been recieved from the api the list will be included in the response to
the client

getGenres - will intiate a call to the movie database api for their full
list of available genres.  Once the list is recieved it will be sent to
the client in the response

getFavorites - will intiate a call to the database for all of the users
favorited movies.  Once the list is recieved it will be sent in the
response to the client

saveMovie - will intiate a call to the database to save the movie that
was sent in the client request to their favorite movie list.  Will
respond to the client once successfully saved

deletMovie - will initiate a call to the database to delete the record
associated with the movie id provided in the client request.  Once it is
deleted a response will be sent to the client