Refactor Movies component into two components
Change Movie componenet to MovieList - renders and <ul> element that
takes the current list of movies and returns a MovieListEntry component
for each of them using the .map method.

MovieListEntry - returns a line entry to be appended to the unordered
list in MovieList component.  Takes in a movie object and displays some
of the elements of the movie on the returned elements.  Attaches methods
to the entries that are passed down from the App component and handle
adding or deleting a movie from the users favorite list.